Initialize ecmult callback data in test_ecmult_multi_batch_single
This fixes a "maybe uninitialized" warning in GCC. The warning is
"semi-spurious": If our code is correct, then ecmult_multi fails due to
an empty scratch space, and the uninitialized values will never be
accessed. But the purpose of this test is to check whether ecmult_multi
really fails when given an empty scratch space, and the correctness of
the test itself should not rely on the correctness of the tested code.