fixup! silentpayments: sending
Renaming this variable is more readable/self-documenting, and I think
it's a good idea here also because we return early if "ret" instead of
"!ret" which is used everywhere else (where ret is a real "success"
code and not some simple boolean return value).

We could also open a new scope and declare sum_is_zero there, but I
didn't do it because it doesn't fit the style of the surrounding code.